+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @constela/cli
 
+## 0.3.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @constela/start@1.2.4
+
 ## 0.3.10
 
 ### Patch Changes